feature #19 : facade 도입
